Revert "Contrastchecker (#1)"
This reverts commit 6c1c4fa3864414d0127da7ed6309f60eb2ec9989.
