+++ src/main.js
@@ -1,9 +1,9 @@
 // import './app.css'
 import 'uno.css';
-import App from './App.svelte';
+import App from './App.svelte'
 
 const app = new App({
   target: document.getElementById('app'),
 })
 
-export default app;
+export default app
