--- src/pages/NotFound.svelte
@@ -1,9 +1,0 @@
-<!-- src/pages/NotFound.svelte -->
-<script>
-  // You can add some logic here if needed
-</script>
-
-<div>
-  <h1>404 - Not Found</h1>
-  <p>The page you are looking for doesn't exist.</p>
-</div>
